Add CSS, improve alert message and popup
Added CSS to improve the popup that is seen by users, and changed the
popup from being an alert to know showing up temporarily in a textbox.
Fixed the error where if user clicked "select answer order", and then
random, the box to input user text would stay on the screen.